# express/server/common/swagger.json
{
  "components": {
    "examples": {},
    "headers": {},
    "parameters": {},
    "requestBodies": {},
    "responses": {},
    "schemas": {},
    "securitySchemes": {
      "api_key": {
        "type": "apiKey",
        "name": "access_token",
        "in": "query"
      },
      "jwt": {
        "type": "apiKey",
        "name": "app_token",
        "in": "header"
      }
    }
  },
  "info": {
    "title": "smart-post-box-express",
    "version": "1.0.0"
  },
  "openapi": "3.0.0",
  "paths": {
    "/photos/last-title/{uniqueCode}": {
      "get": {
        "operationId": "GetLastPhotoTitleByMachineId",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Ok"
          }
        },
        "security": [],
        "parameters": [
          {
            "in": "path",
            "name": "uniqueCode",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/photos/receiver/{uniqueCode}/{index}/{length}/*": {
      "post": {
        "operationId": "ReceivePhotoData",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Ok"
          }
        },
        "security": [],
        "parameters": [
          {
            "in": "path",
            "name": "uniqueCode",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "index",
            "required": true,
            "schema": {
              "format": "double",
              "type": "number"
            }
          },
          {
            "in": "path",
            "name": "length",
            "required": true,
            "schema": {
              "format": "double",
              "type": "number"
            }
          }
        ]
      }
    },
    "/post-box/receiver/{uniqueCode}": {
      "get": {
        "operationId": "ReceiveNotification",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Ok"
          }
        },
        "security": [],
        "parameters": [
          {
            "in": "path",
            "name": "uniqueCode",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/post-box/sync/{uniqueCode}/{modelName}": {
      "post": {
        "operationId": "Sync",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Ok"
          }
        },
        "security": [],
        "parameters": [
          {
            "in": "path",
            "name": "uniqueCode",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "modelName",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/users": {
      "post": {
        "operationId": "GetUsersList",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "properties": {
                      "name": {
                        "type": "string"
                      },
                      "iconUrl": {
                        "type": "string"
                      }
                    },
                    "required": [
                      "iconUrl",
                      "name"
                    ],
                    "type": "object"
                  },
                  "type": "array"
                }
              }
            },
            "description": "Ok"
          }
        },
        "security": [],
        "parameters": []
      }
    }
  },
  "servers": [
    {
      "url": "https://localhost:3000/api"
    }
  ]
}